+++ src/fides/api/tasks/storage.py
@@ -12,7 +12,7 @@ from loguru import logger
 from fides.api.common_exceptions import StorageUploadError
 from fides.api.schemas.storage.storage import ResponseFormat, StorageSecrets
 from fides.api.service.privacy_request.dsr_package.dsr_report_builder import (
-    DsrReportBuilder,
+    DSRReportBuilder,
 )
 from fides.api.service.storage.gcs import get_gcs_blob
 from fides.api.service.storage.s3 import (
@@ -47,7 +47,7 @@ def write_to_in_memory_buffer(
     logger.debug("Writing data to in-memory buffer")
     try:
         if resp_format == ResponseFormat.html.value:
-            return DsrReportBuilder(
+            return DSRReportBuilder(
                 privacy_request=privacy_request,
                 dsr_data=data,
             ).generate()
