+++ tests/ops/service/privacy_request/test_dsr_report_builder.py
@@ -10,21 +10,9 @@ from unittest.mock import patch
 
 import pytest
 
-from fides.api.models.connectionconfig import (
-    AccessLevel,
-    ConnectionConfig,
-    ConnectionType,
-)
-from fides.api.models.datasetconfig import DatasetConfig
 from fides.api.models.privacy_request import PrivacyRequest
-from fides.api.models.privacy_request_redaction_patterns import (
-    PrivacyRequestRedactionPatterns,
-)
-from fides.api.models.sql_models import Dataset as CtlDataset
 from fides.api.service.privacy_request.dsr_package.dsr_report_builder import (
-    DsrReportBuilder,
-    get_redaction_entities_map,
-    get_redaction_entities_map_db,
+    DSRReportBuilder,
 )
 
 
@@ -202,7 +190,7 @@ class TestDsrReportBuilderAttachments(TestDsrReportBuilderBase):
             ],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -255,7 +243,7 @@ class TestDsrReportBuilderAttachments(TestDsrReportBuilderBase):
             "attachments": [webhook_variants["with_attachments"]["attachments"][0]],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -305,7 +293,7 @@ class TestDsrReportBuilderAttachments(TestDsrReportBuilderBase):
             "attachments": [],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -345,7 +333,7 @@ class TestDsrReportBuilderDataStructure:
             "dataset2:collection1": [{"id": 4, "name": "Item 4"}],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -416,7 +404,7 @@ class TestDsrReportBuilderDataStructure:
         }
 
         # Build DSR report - redaction patterns are automatically loaded from privacy request's session
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         # Save the DSR package to a file for inspection
@@ -505,7 +493,7 @@ class TestDsrReportBuilderDataTypes(TestDsrReportBuilderBase):
             ]
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -554,7 +542,7 @@ class TestDsrReportBuilderDataTypes(TestDsrReportBuilderBase):
             ]
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -635,7 +623,7 @@ class TestDsrReportBuilder(TestDsrReportBuilderBase):
         self, privacy_request, dsr_data: dict, common_assertions
     ):
         """Test that the generated report has the correct structure"""
-        builder = DsrReportBuilder(privacy_request, dsr_data)
+        builder = DSRReportBuilder(privacy_request, dsr_data)
         report = builder.generate()
 
         # Verify it's a valid zip file
@@ -654,7 +642,7 @@ class TestDsrReportBuilder(TestDsrReportBuilderBase):
 
     def test_report_content(self, privacy_request, dsr_data: dict):
         """Test that the report content is correctly formatted"""
-        builder = DsrReportBuilder(privacy_request, dsr_data)
+        builder = DSRReportBuilder(privacy_request, dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(report) as zip_file:
@@ -683,7 +671,7 @@ class TestDsrReportBuilder(TestDsrReportBuilderBase):
 
     def test_empty_dsr_data(self, privacy_request, common_assertions):
         """Test report generation with empty DSR data"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         report = builder.generate()
 
         with zipfile.ZipFile(report) as zip_file:
@@ -699,7 +687,7 @@ class TestDsrReportBuilder(TestDsrReportBuilderBase):
 
     def test_privacy_request_mapping(self, privacy_request, dsr_data: dict):
         """Test that privacy request data is correctly mapped in the report"""
-        builder = DsrReportBuilder(privacy_request, dsr_data)
+        builder = DSRReportBuilder(privacy_request, dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(report) as zip_file:
@@ -713,7 +701,7 @@ class TestDsrReportBuilder(TestDsrReportBuilderBase):
 
     def test_attachment_links_in_index(self, privacy_request: PrivacyRequest):
         """Test that attachment links in the index page correctly match the files"""
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data={})
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data={})
 
         # Create multiple attachments with the same name but different URLs
         attachments = [
@@ -765,7 +753,7 @@ class TestDsrReportBuilderAttachmentHandling:
 
     def test_handle_attachment_text(self, privacy_request: PrivacyRequest):
         """Test handling of text attachments with download URLs"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {
                 "file_name": "test.txt",
@@ -782,7 +770,7 @@ class TestDsrReportBuilderAttachmentHandling:
 
     def test_handle_attachment_binary(self, privacy_request: PrivacyRequest):
         """Test handling of binary attachments with download URLs"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {
                 "file_name": "test.pdf",
@@ -799,7 +787,7 @@ class TestDsrReportBuilderAttachmentHandling:
 
     def test_handle_attachment_no_content(self, privacy_request: PrivacyRequest):
         """Test handling of attachments with missing content"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {"file_name": "test.txt", "download_url": None, "file_size": None}
         ]
@@ -823,7 +811,7 @@ class TestDsrReportBuilderAttachmentHandling:
         ]
 
         dsr_data = {"attachments": attachments}
-        builder = DsrReportBuilder(privacy_request, dsr_data)
+        builder = DSRReportBuilder(privacy_request, dsr_data)
         zip_file = builder.generate()
 
         # Create a ZipFile object from the BytesIO
@@ -863,7 +851,7 @@ class TestDsrReportBuilderAttachmentHandling:
             ]
         }
 
-        builder = DsrReportBuilder(privacy_request, dsr_data)
+        builder = DSRReportBuilder(privacy_request, dsr_data)
         zip_file = builder.generate()
 
         # Create a ZipFile object from the BytesIO
@@ -883,7 +871,7 @@ class TestDsrReportBuilderAttachmentHandling:
 
     def test_handle_duplicate_filenames(self, privacy_request: PrivacyRequest):
         """Test handling of duplicate filenames across all directories"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {
                 "file_name": "test.txt",
@@ -916,7 +904,7 @@ class TestDsrReportBuilderAttachmentHandling:
         self, privacy_request: PrivacyRequest
     ):
         """Test handling of duplicate filenames across different directories"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
 
         # Test attachments in different directories
         attachments1 = [
@@ -952,13 +940,13 @@ class TestDsrReportBuilderDatasetHandling:
 
     def test_add_dataset(self, privacy_request: PrivacyRequest):
         """Test adding a dataset with collections"""
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data={})
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data={})
         collections = {
             "collection1": [{"id": 1}, {"id": 2}],
             "collection2": [{"id": 3}],
         }
 
-        builder._add_dataset("dataset1", collections, 1)
+        builder._add_dataset("dataset1", collections)
         builder.out.close()
 
         with zipfile.ZipFile(io.BytesIO(builder.baos.getvalue())) as zip_file:
@@ -974,9 +962,9 @@ class TestDsrReportBuilderDatasetHandling:
 
     def test_add_dataset_empty_collections(self, privacy_request: PrivacyRequest):
         """Test adding a dataset with no collections"""
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data={})
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data={})
 
-        builder._add_dataset("dataset1", {}, 1)
+        builder._add_dataset("dataset1", {})
         builder.out.close()
 
         with zipfile.ZipFile(io.BytesIO(builder.baos.getvalue())) as zip_file:
@@ -1006,7 +994,7 @@ class TestDsrReportBuilderAttachmentContentWriting:
         expected_size,
     ):
         """Test attachment content writing with various file types and sizes"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {
                 "file_name": file_name,
@@ -1048,7 +1036,7 @@ class TestDsrReportBuilderAttachmentContentWriting:
         expected_size,
     ):
         """Test attachment content writing with missing required fields"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {
                 "file_name": file_name,
@@ -1073,7 +1061,7 @@ class TestDsrReportBuilderAttachmentContentWriting:
         self, privacy_request: PrivacyRequest, directory
     ):
         """Test attachment content writing in different directories"""
-        builder = DsrReportBuilder(privacy_request, {})
+        builder = DSRReportBuilder(privacy_request, {})
         attachments = [
             {
                 "file_name": "test.txt",
@@ -1105,7 +1093,7 @@ class TestDsrReportBuilderContent(TestDsrReportBuilderBase):
             "attachments": [],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -1127,7 +1115,7 @@ class TestDsrReportBuilderContent(TestDsrReportBuilderBase):
 
     def test_template_rendering_edge_cases(self, privacy_request: PrivacyRequest):
         """Test template rendering with various edge cases"""
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data={})
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data={})
 
         # Test with missing data
         content = builder._populate_template(
@@ -1156,7 +1144,7 @@ class TestDsrReportBuilderContent(TestDsrReportBuilderBase):
     def test_edge_cases(self, privacy_request: PrivacyRequest):
         """Test various edge cases in the DSR report builder"""
         # Test with empty file names
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data={})
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data={})
         builder._add_file("", "content")  # Should be ignored
         builder._add_file("test.txt", "")  # Should be ignored
 
@@ -1169,7 +1157,7 @@ class TestDsrReportBuilderContent(TestDsrReportBuilderBase):
 
         # Test with empty attachments
         dsr_data = {"dataset1:collection1": [{"id": 1, "attachments": []}]}
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -1198,7 +1186,7 @@ class TestDsrReportBuilderOrganization(TestDsrReportBuilderBase):
             "charlie:collection1": [{"id": 5}],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -1261,7 +1249,7 @@ class TestDsrReportBuilderOrganization(TestDsrReportBuilderBase):
             "dataset2:collection1": [{"id": 4, "name": "Item 4"}],
         }
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
 
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
@@ -1293,7 +1281,7 @@ class TestDsrReportBuilderOrganization(TestDsrReportBuilderBase):
     def test_concurrent_access(self, privacy_request: PrivacyRequest):
         """Test concurrent access to the zip file"""
 
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data={})
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data={})
         errors = []
 
         def write_file():
@@ -1331,7 +1319,7 @@ class TestDsrReportBuilderOrganization(TestDsrReportBuilderBase):
         }
 
         start_time = time.time()
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
+        builder = DSRReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
         report = builder.generate()
         end_time = time.time()
 
@@ -1343,1102 +1331,3 @@ class TestDsrReportBuilderOrganization(TestDsrReportBuilderBase):
         with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
             assert "data/dataset1/index.html" in zip_file.namelist()
             assert len([f for f in zip_file.namelist() if "collection" in f]) == 100
-
-
-@pytest.mark.integration_postgres
-class TestGetRedactionEntitiesMap:
-    """Tests for both get_redaction_entities_map implementations (Python and DB-based)"""
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_empty_dataset(self, db, redaction_func):
-        """Test get_redaction_entities_map with no dataset configurations"""
-        result = redaction_func(db)
-        assert result == set()
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_no_redaction_meta(
-        self, db, postgres_example_test_dataset_config, redaction_func
-    ):
-        """Test get_redaction_entities_map with dataset config but no redaction meta"""
-        result = redaction_func(db)
-        assert result == set()
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_dataset_level_redaction(
-        self, db, redaction_func
-    ):
-        """Test get_redaction_entities_map with dataset-level redaction"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_redaction_connection",
-                "name": "Test Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with dataset-level redaction
-        dataset_dict = {
-            "fides_key": "test_dataset_redacted",
-            "name": "test_dataset_redacted",
-            "collections": [],
-            "fides_meta": {"redact": "name"},
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "test_dataset_redacted",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        result = redaction_func(db)
-        assert "test_dataset_redacted" in result
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_collection_level_redaction(
-        self, db, redaction_func
-    ):
-        """Test get_redaction_entities_map with collection-level redaction"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_collection_redaction_connection",
-                "name": "Test Collection Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with collection-level redaction
-        dataset_dict = {
-            "fides_key": "test_dataset",
-            "name": "test_dataset",
-            "collections": [
-                {
-                    "name": "users",
-                    "fields": [],
-                    "fides_meta": {"redact": "name"},
-                },
-                {
-                    "name": "orders",
-                    "fields": [],
-                    # No redaction meta
-                },
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "test_dataset",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        result = redaction_func(db)
-        assert "test_dataset.users" in result
-        assert "test_dataset.orders" not in result
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_field_level_redaction(self, db, redaction_func):
-        """Test get_redaction_entities_map with field-level redaction"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_field_redaction_connection",
-                "name": "Test Field Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with field-level redaction
-        dataset_dict = {
-            "fides_key": "test_dataset",
-            "name": "test_dataset",
-            "collections": [
-                {
-                    "name": "users",
-                    "fields": [
-                        {
-                            "name": "email",
-                            "data_categories": ["user.contact.email"],
-                            "fides_meta": {"redact": "name"},
-                        },
-                        {
-                            "name": "name",
-                            "data_categories": ["user.name"],
-                            # No redaction meta
-                        },
-                    ],
-                }
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "test_dataset",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        result = redaction_func(db)
-        assert "test_dataset.users.email" in result
-        assert "test_dataset.users.name" not in result
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_hierarchical_redaction(
-        self, db, redaction_func
-    ):
-        """Test get_redaction_entities_map with redaction at multiple levels"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_hierarchical_redaction_connection",
-                "name": "Test Hierarchical Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with redaction at dataset, collection, and field levels
-        dataset_dict = {
-            "fides_key": "test_dataset",
-            "name": "test_dataset",
-            "fides_meta": {"redact": "name"},  # Dataset level
-            "collections": [
-                {
-                    "name": "users",
-                    "fides_meta": {"redact": "name"},  # Collection level
-                    "fields": [
-                        {
-                            "name": "email",
-                            "data_categories": ["user.contact.email"],
-                            "fides_meta": {"redact": "name"},  # Field level
-                        },
-                        {
-                            "name": "phone",
-                            "data_categories": ["user.contact.phone_number"],
-                            # No redaction meta
-                        },
-                    ],
-                },
-                {
-                    "name": "orders",
-                    "fields": [
-                        {
-                            "name": "order_id",
-                            "data_categories": ["system.operations"],
-                            "fides_meta": {"redact": "name"},  # Field level only
-                        },
-                    ],
-                },
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "test_dataset",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        result = redaction_func(db)
-        # Dataset level
-        assert "test_dataset" in result
-        # Collection level
-        assert "test_dataset.users" in result
-        assert "test_dataset.orders" not in result
-        # Field level
-        assert "test_dataset.users.email" in result
-        assert "test_dataset.users.phone" not in result
-        assert "test_dataset.orders.order_id" in result
-
-    def test_nested_field_regex_pattern_redaction(
-        self, privacy_request: PrivacyRequest, db
-    ):
-        """Test that nested JSON fields get redacted when regex pattern is configured."""
-        # Set up redaction patterns
-        PrivacyRequestRedactionPatterns.create_or_update(
-            db=db, data={"patterns": ["full_name", "email.*"]}
-        )
-
-        # Create connection config
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_nested_redaction",
-                "name": "Test Nested Redaction",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create dataset with JSON field
-        dataset_dict = {
-            "fides_key": "nested_test",
-            "name": "nested_test",
-            "collections": [
-                {
-                    "name": "profiles",
-                    "fields": [
-                        {"name": "id", "data_categories": ["system.operations"]},
-                        {"name": "user_info", "data_categories": ["user"]},
-                    ],
-                }
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "nested_test",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        # Create DSR data with nested JSON
-        dsr_data = {
-            "nested_test:profiles": [
-                {
-                    "id": 1,
-                    "user_info": {
-                        "full_name": "John Doe",
-                        "email_address": "john@example.com",
-                        "phone": "555-1234",
-                    },
-                },
-            ],
-        }
-
-        # Generate report
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        # Check redaction
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            content = zip_file.read("data/nested_test/profiles/index.html").decode(
-                "utf-8"
-            )
-
-            # Nested fields matching regex patterns should be redacted
-            assert "full_name" not in content
-            assert "email_address" not in content
-            # Non-matching fields should remain
-            assert "phone" in content
-            # Data values should be preserved
-            assert "John Doe" in content
-            assert "john@example.com" in content
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_multiple_datasets(self, db, redaction_func):
-        """Test get_redaction_entities_map with multiple datasets"""
-
-        connection_configs = []
-
-        # Create multiple datasets with different redaction configurations
-        for dataset_index in range(1, 4):
-            # Create a connection config for each dataset
-            connection_config = ConnectionConfig.create(
-                db=db,
-                data={
-                    "key": f"test_multiple_datasets_connection_{dataset_index}",
-                    "name": f"Test Multiple Datasets Connection {dataset_index}",
-                    "connection_type": ConnectionType.manual,
-                    "access": AccessLevel.read,
-                    "disabled": False,
-                },
-            )
-            connection_configs.append(connection_config)
-
-            dataset_dict = {
-                "fides_key": f"dataset_{dataset_index}",
-                "name": f"dataset_{dataset_index}",
-                "collections": [
-                    {
-                        "name": "users",
-                        "fields": [
-                            {
-                                "name": "email",
-                                "data_categories": ["user.contact.email"],
-                                "fides_meta": (
-                                    {"redact": "name"} if dataset_index % 2 == 1 else {}
-                                ),
-                            },
-                        ],
-                        "fides_meta": {"redact": "name"} if dataset_index == 2 else {},
-                    }
-                ],
-                "fides_meta": {"redact": "name"} if dataset_index == 3 else {},
-            }
-
-            ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-            DatasetConfig.create(
-                db=db,
-                data={
-                    "connection_config_id": connection_config.id,
-                    "fides_key": f"dataset_{dataset_index}",
-                    "ctl_dataset_id": ctl_dataset.id,
-                },
-            )
-
-        result = redaction_func(db)
-
-        # Dataset 1: field-level redaction only
-        assert "dataset_1.users.email" in result
-        assert "dataset_1.users" not in result
-        assert "dataset_1" not in result
-
-        # Dataset 2: collection-level redaction
-        assert "dataset_2.users" in result
-        assert "dataset_2" not in result
-
-        # Dataset 3: dataset-level redaction
-        assert "dataset_3" in result
-
-    @pytest.mark.parametrize(
-        "redaction_func",
-        [get_redaction_entities_map, get_redaction_entities_map_db],
-        ids=["python_implementation", "db_implementation"],
-    )
-    def test_get_redaction_entities_map_nested_fields(self, db, redaction_func):
-        """Test get_redaction_entities_map with deeply nested fields"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_nested_connection",
-                "name": "Test Nested Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with deeply nested fields
-        dataset_dict = {
-            "fides_key": "test_nested_dataset",
-            "name": "test_nested_dataset",
-            "collections": [
-                {
-                    "name": "users",
-                    "fields": [
-                        {
-                            "name": "profile",
-                            "fides_meta": {"redact": "name"},  # Level 1 redaction
-                            "fields": [
-                                {
-                                    "name": "personal_info",
-                                    "fides_meta": {
-                                        "redact": "name"
-                                    },  # Level 2 redaction
-                                    "fields": [
-                                        {
-                                            "name": "ssn",
-                                            "fides_meta": {
-                                                "redact": "name"
-                                            },  # Level 3 redaction
-                                            "fields": [],
-                                        },
-                                        {
-                                            "name": "passport_info",
-                                            "fides_meta": {},  # No redaction
-                                            "fields": [
-                                                {
-                                                    "name": "passport_number",
-                                                    "fides_meta": {
-                                                        "redact": "name"
-                                                    },  # Level 4 redaction
-                                                    "fields": [],
-                                                }
-                                            ],
-                                        },
-                                    ],
-                                }
-                            ],
-                        },
-                        {
-                            "name": "email",
-                            "fides_meta": {},  # No redaction
-                            "fields": [],
-                        },
-                    ],
-                }
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "test_nested_dataset",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        result = redaction_func(db)
-
-        # Verify all nested redaction entities are found
-        assert "test_nested_dataset.users.profile" in result
-        assert "test_nested_dataset.users.profile.personal_info" in result
-        assert "test_nested_dataset.users.profile.personal_info.ssn" in result
-        assert (
-            "test_nested_dataset.users.profile.personal_info.passport_info.passport_number"
-            in result
-        )
-
-        # Verify non-redacted entities are not included
-        assert "test_nested_dataset.users.email" not in result
-        assert (
-            "test_nested_dataset.users.profile.personal_info.passport_info"
-            not in result
-        )
-
-    def test_both_implementations_return_same_results(self, db):
-        """Test that both implementations return identical results for complex dataset"""
-
-        # Create a connection config
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_comparison_connection",
-                "name": "Test Comparison Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a complex dataset with redaction at all levels
-        dataset_dict = {
-            "fides_key": "comparison_dataset",
-            "name": "comparison_dataset",
-            "fides_meta": {"redact": "name"},
-            "collections": [
-                {
-                    "name": "collection_one",
-                    "fields": [
-                        {
-                            "name": "field_one",
-                            "data_categories": ["user"],
-                            "fides_meta": {"redact": "name"},
-                        },
-                        {
-                            "name": "field_two",
-                            "data_categories": ["user"],
-                            "fields": [
-                                {
-                                    "name": "nested_field",
-                                    "data_categories": ["user"],
-                                    "fides_meta": {"redact": "name"},
-                                }
-                            ],
-                        },
-                    ],
-                    "fides_meta": {"redact": "name"},
-                },
-                {
-                    "name": "collection_two",
-                    "fields": [
-                        {
-                            "name": "another_field",
-                            "data_categories": ["user"],
-                        }
-                    ],
-                },
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "comparison_dataset",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        # Get results from both implementations
-        python_result = get_redaction_entities_map(db)
-        db_result = get_redaction_entities_map_db(db)
-
-        # Both should return the same entities
-        assert python_result == db_result
-
-        # Verify specific entities are present in both
-        expected_entities = {
-            "comparison_dataset",  # Dataset level
-            "comparison_dataset.collection_one",  # Collection level
-            "comparison_dataset.collection_one.field_one",  # Field level
-            "comparison_dataset.collection_one.field_two.nested_field",  # Nested field
-        }
-
-        for entity in expected_entities:
-            assert entity in python_result
-            assert entity in db_result
-
-
-@pytest.mark.integration_postgres
-class TestDsrReportBuilderRedactionEntitiesIntegration(TestDsrReportBuilderBase):
-    """Tests for DSR report builder integration with redaction entities map"""
-
-    def test_redaction_by_entities_map_dataset_level(
-        self, privacy_request: PrivacyRequest, db
-    ):
-        """Test dataset-level redaction using entities map"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_integration_dataset_redaction_connection",
-                "name": "Test Integration Dataset Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with dataset-level redaction
-        dataset_dict = {
-            "fides_key": "customer_database",
-            "name": "customer_database",
-            "collections": [],
-            "fides_meta": {"redact": "name"},
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "customer_database",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        dsr_data = {
-            "customer_database:users": [
-                {"id": 1, "name": "John Doe", "email": "john@example.com"},
-            ],
-            "public_data:articles": [
-                {"id": 1, "title": "Public Article"},
-            ],
-        }
-
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            welcome_content = zip_file.read("welcome.html").decode("utf-8")
-
-            # customer_database should be redacted to dataset_1
-            assert "dataset_1" in welcome_content
-            assert "customer_database" not in welcome_content
-
-            # public_data should not be redacted (no redaction configuration)
-            assert "public_data" in welcome_content
-
-            # Verify file structure uses redacted names
-            assert "data/dataset_1/index.html" in zip_file.namelist()
-            assert "data/public_data/index.html" in zip_file.namelist()
-
-    def test_redaction_by_entities_map_collection_level(
-        self, privacy_request: PrivacyRequest, db
-    ):
-        """Test collection-level redaction using entities map"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_integration_collection_redaction_connection",
-                "name": "Test Integration Collection Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with collection-level redaction
-        dataset_dict = {
-            "fides_key": "customer_data",
-            "name": "customer_data",
-            "collections": [
-                {
-                    "name": "user_profiles",
-                    "fields": [],
-                    "fides_meta": {"redact": "name"},
-                },
-                {
-                    "name": "orders",
-                    "fields": [],
-                    # No redaction meta
-                },
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "customer_data",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        dsr_data = {
-            "customer_data:user_profiles": [
-                {"id": 1, "name": "John Doe"},
-            ],
-            "customer_data:orders": [
-                {"id": 101, "amount": 49.99},
-            ],
-        }
-
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            dataset_content = zip_file.read("data/customer_data/index.html").decode(
-                "utf-8"
-            )
-
-            # user_profiles should be redacted to collection_1
-            assert "collection_1" in dataset_content
-            assert "user_profiles" not in dataset_content
-
-            # orders should not be redacted
-            assert "orders" in dataset_content
-
-            # Verify file structure
-            assert "data/customer_data/collection_1/index.html" in zip_file.namelist()
-            assert "data/customer_data/orders/index.html" in zip_file.namelist()
-
-    def test_redaction_by_entities_map_field_level(
-        self, privacy_request: PrivacyRequest, db
-    ):
-        """Test field-level redaction using entities map"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_integration_field_redaction_connection",
-                "name": "Test Integration Field Redaction Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create a dataset with field-level redaction
-        dataset_dict = {
-            "fides_key": "user_data",
-            "name": "user_data",
-            "collections": [
-                {
-                    "name": "profiles",
-                    "fields": [
-                        {
-                            "name": "email_address",
-                            "data_categories": ["user.contact.email"],
-                            "fides_meta": {"redact": "name"},
-                        },
-                        {
-                            "name": "full_name",
-                            "data_categories": ["user.name"],
-                            # No redaction meta
-                        },
-                    ],
-                }
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "user_data",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        dsr_data = {
-            "user_data:profiles": [
-                {"id": 1, "email_address": "john@example.com", "full_name": "John Doe"},
-            ],
-        }
-
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            collection_content = zip_file.read(
-                "data/user_data/profiles/index.html"
-            ).decode("utf-8")
-
-            # email_address should be redacted to field_N
-            assert "field_" in collection_content
-            assert "email_address" not in collection_content
-
-            # full_name should not be redacted
-            assert "full_name" in collection_content
-
-            # Verify the data values are still present
-            assert "john@example.com" in collection_content
-            assert "John Doe" in collection_content
-
-    def test_redaction_entities_map_priority_over_regex(
-        self, privacy_request: PrivacyRequest, db
-    ):
-        """Test that entities map redaction takes priority over regex patterns"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_integration_regex_priority_connection",
-                "name": "Test Integration Regex Priority Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create dataset config with specific redaction for one collection
-        dataset_dict = {
-            "fides_key": "test_data",
-            "name": "test_data",
-            "collections": [
-                {
-                    "name": "user_accounts",  # Would match regex .*user.* but has specific redaction
-                    "fields": [],
-                    "fides_meta": {"redact": "name"},
-                },
-                {
-                    "name": "user_sessions",  # Would match regex .*user.* but no specific redaction
-                    "fields": [],
-                },
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "test_data",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        # Set up regex patterns that would match both collections
-        PrivacyRequestRedactionPatterns.create_or_update(
-            db=db, data={"patterns": [r".*user.*"]}
-        )
-
-        dsr_data = {
-            "test_data:user_accounts": [{"id": 1}],
-            "test_data:user_sessions": [{"id": 2}],
-        }
-
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            dataset_content = zip_file.read("data/test_data/index.html").decode("utf-8")
-
-            # user_accounts has specific redaction config, so uses collection_1
-            assert "collection_1" in dataset_content
-
-            # user_sessions has no specific config but matches regex, so uses collection_2
-            assert "collection_2" in dataset_content
-
-            # Neither original name should appear
-            assert "user_accounts" not in dataset_content
-            assert "user_sessions" not in dataset_content
-
-    def test_redaction_entities_map_with_mixed_hierarchical_levels(
-        self, privacy_request: PrivacyRequest, db
-    ):
-        """Test redaction with entities at different hierarchical levels including nested data structures"""
-
-        # Create a connection config first (ensure it's not disabled)
-        connection_config = ConnectionConfig.create(
-            db=db,
-            data={
-                "key": "test_integration_mixed_hierarchical_connection",
-                "name": "Test Integration Mixed Hierarchical Connection",
-                "connection_type": ConnectionType.manual,
-                "access": AccessLevel.read,
-                "disabled": False,
-            },
-        )
-
-        # Create dataset with redaction at multiple levels including nested fields
-        dataset_dict = {
-            "fides_key": "mixed_data",
-            "name": "mixed_data",
-            "fides_meta": {"redact": "name"},  # Dataset level
-            "collections": [
-                {
-                    "name": "public_info",
-                    "fields": [
-                        {
-                            "name": "description",
-                            "data_categories": ["system.operations"],
-                        },
-                        {
-                            "name": "metadata",
-                            "data_categories": ["system.operations"],
-                            "fields": [
-                                {
-                                    "name": "created_by",
-                                    "data_categories": ["user.name"],
-                                    "fides_meta": {"redact": "name"},
-                                },
-                                {
-                                    "name": "tags",
-                                    "data_categories": ["system.operations"],
-                                    "fields": [
-                                        {
-                                            "name": "category",
-                                            "data_categories": ["system.operations"],
-                                        }
-                                    ],
-                                },
-                            ],
-                        },
-                    ],
-                    # No collection-level redaction (should inherit from dataset)
-                },
-                {
-                    "name": "sensitive_data",
-                    "fields": [
-                        {
-                            "name": "ssn",
-                            "data_categories": ["user.government_id"],
-                            "fides_meta": {
-                                "redact": "name"
-                            },  # Field level (redundant with dataset)
-                        },
-                        {
-                            "name": "personal_details",
-                            "data_categories": ["user"],
-                            "fields": [
-                                {
-                                    "name": "address",
-                                    "data_categories": ["user.contact.address"],
-                                    "fides_meta": {"redact": "name"},
-                                    "fields": [
-                                        {
-                                            "name": "street",
-                                            "data_categories": [
-                                                "user.contact.address.street"
-                                            ],
-                                        },
-                                        {
-                                            "name": "postal_code",
-                                            "data_categories": [
-                                                "user.contact.address.postal_code"
-                                            ],
-                                            "fides_meta": {"redact": "name"},
-                                        },
-                                    ],
-                                },
-                                {
-                                    "name": "phone_numbers",
-                                    "data_categories": ["user.contact.phone_number"],
-                                },
-                            ],
-                        },
-                    ],
-                    "fides_meta": {
-                        "redact": "name"
-                    },  # Collection level (redundant with dataset)
-                },
-            ],
-        }
-
-        ctl_dataset = CtlDataset.create_from_dataset_dict(db, dataset_dict)
-
-        DatasetConfig.create(
-            db=db,
-            data={
-                "connection_config_id": connection_config.id,
-                "fides_key": "mixed_data",
-                "ctl_dataset_id": ctl_dataset.id,
-            },
-        )
-
-        dsr_data = {
-            "mixed_data:public_info": [
-                {
-                    "description": "public data",
-                    "metadata": {
-                        "created_by": "john_doe",
-                        "tags": {"category": "general"},
-                    },
-                }
-            ],
-            "mixed_data:sensitive_data": [
-                {
-                    "ssn": "123-45-6789",
-                    "personal_details": {
-                        "address": {"street": "123 Main St", "postal_code": "12345"},
-                        "phone_numbers": ["555-1234", "555-5678"],
-                    },
-                }
-            ],
-        }
-
-        builder = DsrReportBuilder(privacy_request=privacy_request, dsr_data=dsr_data)
-        report = builder.generate()
-
-        with zipfile.ZipFile(io.BytesIO(report.getvalue())) as zip_file:
-            welcome_content = zip_file.read("welcome.html").decode("utf-8")
-
-            # Dataset is redacted due to dataset-level config
-            assert "dataset_1" in welcome_content
-            assert "mixed_data" not in welcome_content
-
-            # Check that file structure uses redacted dataset name
-            assert "data/dataset_1/index.html" in zip_file.namelist()
-
-            # Verify nested data is properly included in the generated report
-            dataset_content = zip_file.read("data/dataset_1/index.html").decode("utf-8")
-
-            # Based on the actual output, public_info is not redacted but sensitive_data is
-            # This is because only sensitive_data has explicit collection-level redaction
-            assert (
-                "public_info" in dataset_content
-            )  # public_info not redacted (no collection-level config)
-            assert (
-                "collection_2" in dataset_content
-            )  # sensitive_data redacted due to collection-level config
-            assert "sensitive_data" not in dataset_content
-
-            # Check that nested data structure is preserved in the report
-            # Look for collection detail pages
-            public_info_files = [
-                f
-                for f in zip_file.namelist()
-                if "public_info" in f and f.endswith(".html")
-            ]
-            sensitive_data_files = [
-                f
-                for f in zip_file.namelist()
-                if "collection_2" in f and f.endswith(".html")
-            ]
-
-            assert (
-                len(public_info_files) > 0
-            ), "Should have public_info collection files"
-            assert (
-                len(sensitive_data_files) > 0
-            ), "Should have sensitive_data collection files"
-
-            # Verify nested data appears in the collection content
-            if public_info_files:
-                public_content = zip_file.read(public_info_files[0]).decode("utf-8")
-
-                # Verify that nested field names are redacted in the JSON content
-                # Only "created_by" has explicit redaction config, so only it should be redacted
-                assert (
-                    "created_by" not in public_content
-                )  # field name should be redacted in JSON
-                assert "field_" in public_content  # should contain redacted field names
-
-                # The "tags" field should NOT be redacted, it has no explicit redaction config
-                assert "tags" in public_content  # tags should remain unredacted
-
-            if sensitive_data_files:
-                sensitive_content = zip_file.read(sensitive_data_files[0]).decode(
-                    "utf-8"
-                )
-
-                # Verify that nested field names with explicit redaction configs are redacted
-                assert (
-                    "address" not in sensitive_content
-                )  # address field has explicit redaction config
-                assert (
-                    "postal_code" not in sensitive_content
-                )  # postal_code field has explicit redaction config
-                assert (
-                    "field_" in sensitive_content
-                )  # should contain redacted field names
-
-                # Fields without explicit redaction should remain unredacted
-                assert (
-                    "street" in sensitive_content
-                )  # street has no explicit redaction config
-                assert (
-                    "phone_numbers" in sensitive_content
-                )  # phone_numbers has no explicit redaction config
